+++ app/src/main/java/nextstep/jwp/request/RequestLine.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.request;
 
 import nextstep.jwp.constants.Http;
 import nextstep.jwp.constants.HttpMethod;
